[IOS-6810] Revert init update, update load and callback methods
This commit reverts previous change to init and instead removes the
VungleAds init check in all ad type load events. This PR should be
reviewed in conjunction with changes in the IOS-6810 branch in the
VungleAdsSDK repo.

By removing the init check, the adapter can all load for all ad types
no matter if the SDK is initialized or not - in the event the SDK init
previously failed, this will allow for the SDK to retry the init as part of
the load event. If the init fails, the ad load will fail with an SDK not
initialized error. If the init passes, the ad load will continue. The adapter
has been updated so that the initialization status can be updated
during the ad load or ad load failure callbacks.

IOS-6810